pipeline-image-classification-file/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pika==1.3.2
pillow==10.2.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.2
sympy==1.12
timm==0.9.16
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
